refactor of aircraft package
