+++ app/src/main/java/org/WenuLink/adapters/aircraft/AircraftHandler.kt
@@ -1,4 +1,4 @@
-package org.WenuLink.adapters
+package org.WenuLink.adapters.aircraft
 
 import com.MAVLink.enums.MAV_MODE_FLAG
 import io.getstream.log.taggedLogger
@@ -10,8 +10,9 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
-import org.WenuLink.adapters.aircraft.AircraftCommand
-import org.WenuLink.adapters.aircraft.BootCommand
+import org.WenuLink.adapters.AsyncUtils
+import org.WenuLink.adapters.CameraHandler
+import org.WenuLink.adapters.MissionHandler
 import org.WenuLink.sdk.FCManager
 import org.WenuLink.sdk.MissionActionManager
 import org.WenuLink.sdk.MissionManager
@@ -45,12 +46,12 @@ class AircraftHandler {
     var homeTimestamp: Long = startTimestamp
     var sensorsHealthy = false
         private set
-    val mission = MissionHandler.getInstance()
+    val mission = MissionHandler.Companion.getInstance()
     val telemetry = TelemetryHandler.getInstance()
     var isPowerOff = true
     val rcInput: RCData?
         get() = telemetry.getRCData()
-    val cameras = CameraHandler.getInstance()
+    val cameras = CameraHandler.Companion.getInstance()
     var hasCameras: Boolean = false
     private var commandJob: Job? = null
     private val commandChannel =
