+++ app/src/main/java/org/WenuLink/controllers/CameraController.kt
@@ -16,13 +16,13 @@ import com.MAVLink.enums.STORAGE_TYPE
 import com.MAVLink.enums.STORAGE_USAGE_FLAG
 import io.getstream.log.taggedLogger
 import kotlin.math.roundToLong
-import org.WenuLink.adapters.AircraftHandler
-import org.WenuLink.adapters.CameraCaptureStatus
-import org.WenuLink.adapters.CameraMetadata
-import org.WenuLink.adapters.ImageMetadata
+import org.WenuLink.adapters.aircraft.AircraftHandler
+import org.WenuLink.adapters.aircraft.CameraCaptureStatus
+import org.WenuLink.adapters.aircraft.CameraMetadata
+import org.WenuLink.adapters.aircraft.ImageMetadata
 import org.WenuLink.adapters.MessageUtils
 import org.WenuLink.adapters.OrientationUtils
-import org.WenuLink.adapters.TelemetryMapper
+import org.WenuLink.adapters.aircraft.TelemetryMapper
 import org.WenuLink.adapters.actions.SetModeCommand
 import org.WenuLink.adapters.actions.StartRecordCommand
 import org.WenuLink.adapters.actions.StopRecordCommand
