+++ app/src/main/java/org/WenuLink/controllers/ConnectionController.kt
@@ -24,9 +24,9 @@ import io.getstream.log.taggedLogger
 import kotlin.math.pow
 import kotlin.math.roundToInt
 import kotlin.math.sqrt
-import org.WenuLink.adapters.AircraftHandler
+import org.WenuLink.adapters.aircraft.AircraftHandler
 import org.WenuLink.adapters.MessageUtils
-import org.WenuLink.adapters.TelemetryHandler
+import org.WenuLink.adapters.aircraft.TelemetryHandler
 import org.WenuLink.mavlink.MAVLinkClient
 
 /**
