+++ app/src/main/java/org/WenuLink/views/ServicesViewModel.kt
@@ -11,9 +11,9 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import org.WenuLink.WenuLinkApp
-import org.WenuLink.adapters.AircraftHandler
+import org.WenuLink.adapters.aircraft.AircraftHandler
 import org.WenuLink.adapters.AsyncUtils
-import org.WenuLink.adapters.TelemetryData
+import org.WenuLink.adapters.aircraft.TelemetryData
 
 class ServicesViewModel(application: Application) : AndroidViewModel(application) {
     private val logger by taggedLogger(ServicesViewModel::class.java.simpleName)
